svc-systems/systems/storage.go
//(C) Copyright [2020] Hewlett Packard Enterprise Development LP
//
//Licensed under the Apache License, Version 2.0 (the "License"); you may
//not use this file except in compliance with the License. You may obtain
//a copy of the License at
//
//    http://www.apache.org/licenses/LICENSE-2.0
//
//Unless required by applicable law or agreed to in writing, software
//distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
//WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
//License for the specific language governing permissions and limitations
// under the License.

//Package systems ...
package systems

import (
	"net/http"

	systemsproto "github.com/ODIM-Project/ODIM/lib-utilities/proto/systems"
	"github.com/ODIM-Project/ODIM/lib-utilities/response"
)

// CreateVolume defines the logic for creating a volume under storage
func (p *PluginContact) CreateVolume(req *systemsproto.CreateVolumeRequest) response.RPC {
	// This function is yet to be implemented
	var resp response.RPC
	resp.StatusCode = http.StatusNotImplemented
	return resp
}
